+++ cpp/oneapi/dal/algo/jaccard/detail/vertex_similarity_default_kernel.hpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright 2020 Intel Corporation
+* Copyright 2020-2021 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -19,35 +19,37 @@
 #include "oneapi/dal/algo/jaccard/common.hpp"
 #include "oneapi/dal/algo/jaccard/vertex_similarity_types.hpp"
 #include "oneapi/dal/graph/detail/undirected_adjacency_vector_graph_impl.hpp"
+#include "oneapi/dal/table/detail/table_builder.hpp"
 
 namespace oneapi::dal::preview::jaccard::detail {
 
-inline std::size_t get_number_elements_in_block(const std::int32_t &row_range_begin,
-                                                const std::int32_t &row_range_end,
-                                                const std::int32_t &column_range_begin,
-                                                const std::int32_t &column_range_end) {
+inline std::int64_t get_number_elements_in_block(const std::int32_t &row_range_begin,
+                                                 const std::int32_t &row_range_end,
+                                                 const std::int32_t &column_range_begin,
+                                                 const std::int32_t &column_range_end) {
     ONEDAL_ASSERT(row_range_end >= row_range_begin, "Negative interval found");
-    const std::size_t row_count = row_range_end - row_range_begin;
+    const std::int64_t row_count = row_range_end - row_range_begin;
     ONEDAL_ASSERT(column_range_end >= column_range_begin, "Negative interval found");
-    const std::size_t column_count = column_range_end - column_range_begin;
+    const std::int64_t column_count = column_range_end - column_range_begin;
     // compute the number of the vertex pairs in the block of the graph
-    const std::size_t vertex_pairs_count = row_count * column_count;
+    const std::int64_t vertex_pairs_count = row_count * column_count;
     ONEDAL_ASSERT(vertex_pairs_count / row_count == column_count,
                   "Overflow found in multiplication of two values");
     return vertex_pairs_count;
 }
 
 template <typename Float, typename Index>
-inline std::size_t get_max_block_size(const std::int64_t &vertex_pairs_count) {
-    const std::size_t vertex_pair_element_count = 2; // 2 elements in the vertex pair
-    const std::size_t jaccard_coeff_element_count = 1; // 1 Jaccard coeff for the vertex pair
+inline std::int64_t get_max_block_size(const std::int64_t &vertex_pairs_count) {
+    const std::int64_t vertex_pair_element_count = 2; // 2 elements in the vertex pair
+    const std::int64_t jaccard_coeff_element_count = 1; // 1 Jaccard coeff for the vertex pair
 
-    const std::size_t vertex_pair_size = vertex_pair_element_count * sizeof(Index); // size in bytes
-    const std::size_t jaccard_coeff_size =
+    const std::int64_t vertex_pair_size =
+        vertex_pair_element_count * sizeof(Index); // size in bytes
+    const std::int64_t jaccard_coeff_size =
         jaccard_coeff_element_count * sizeof(Float); // size in bytes
-    const std::size_t element_result_size = vertex_pair_size + jaccard_coeff_size;
+    const std::int64_t element_result_size = vertex_pair_size + jaccard_coeff_size;
 
-    const std::size_t block_result_size = element_result_size * vertex_pairs_count;
+    const std::int64_t block_result_size = element_result_size * vertex_pairs_count;
     ONEDAL_ASSERT(block_result_size / vertex_pairs_count == element_result_size,
                   "Overflow found in multiplication of two values");
     return block_result_size;
@@ -64,7 +66,7 @@ inline Index max(const Index &a, const Index &b) {
 }
 
 template <typename Index>
-inline std::size_t intersection(const Index *neigh_u, const Index *neigh_v, Index n_u, Index n_v);
+inline std::int64_t intersection(const Index *neigh_u, const Index *neigh_v, Index n_u, Index n_v);
 
 template <typename Index>
 vertex_similarity_result call_jaccard_default_kernel_general(
@@ -141,8 +143,8 @@ vertex_similarity_result call_jaccard_default_kernel_general(
 }
 
 template <typename Index>
-inline std::size_t intersection(const Index *neigh_u, const Index *neigh_v, Index n_u, Index n_v) {
-    std::size_t total = 0;
+inline std::int64_t intersection(const Index *neigh_u, const Index *neigh_v, Index n_u, Index n_v) {
+    std::int64_t total = 0;
     Index i_u = 0, i_v = 0;
     while (i_u < n_u && i_v < n_v) {
         if ((neigh_u[i_u] > neigh_v[n_v - 1]) || (neigh_v[i_v] > neigh_u[n_u - 1])) {
